+++ scripts/asset-lookup.js
@@ -14,7 +14,7 @@ import {
     dataVolume,
     dataBanners,
     dataBackgrounds
-} from './data-collections.js';
+} from '../resources/assets.js';
 
 export class AssetLookup {