+++ scripts/settings.js
@@ -10,7 +10,7 @@ import { MODULE, BLACKSMITH } from './const.js';
 import { registerBlacksmithUpdatedHook, postConsoleAndNotification, getActorId, resetModuleSettings, getSettingSafely, setSettingSafely } from './api-common.js';
 // -- Import special page variables --
 // Load the data sets for the settings dropdowns
-import { dataNameplate, dataSounds, dataIcons, dataBackgroundImages, dataTheme } from './data-collections.js';
+import { dataNameplate, dataSounds, dataIcons, dataBackgroundImages, dataTheme } from '../resources/assets.js';